Landing page loads after file move, need to fix CSS import
