+++ src/components/LandingPage/app.js
@@ -5,7 +5,7 @@ import { Route, withRouter, } from 'react-router-dom';
 import HeaderBar from './header-bar';
 import LandingPage from './Landing-Page';
 
-import Application from './applicationPage/Application';
+import Application from '.././applicationPage/Application';
 import RegistrationPage from './registration-page';
 import { refreshAuthToken, } from '../../actions/auth';
